+++ src/main/kotlin/blackjack/README.md
@@ -1,0 +1,12 @@
+## 블랙잭 과제
+
+### 기능 명세
+- [x] 카드의 숫자 계산은 카드 숫자를 기본으로 하며, 예외로 Ace는 1 또는 11로 계산할 수 있으며, King, Queen, Jack은 각각 10으로 계산한다.
+- [x] 게임을 시작하면 플레이어는 두 장의 카드를 지급 받으며, 두 장의 카드 숫자를 합쳐 21을 초과하지 않으면서 21에 가깝게 만들면 이긴다. 21을 넘지 않을 경우 원한다면 얼마든지 카드를 계속 뽑을 수 있다.
+
+### 리뷰 받은 것
+- [ ] 일반적으로 Deck은 전체 카드 뭉치를 의미합니다. 플레이어의 손패는 Hand라고 하구요.
+- [ ] 이렇게 되면 카드가 중복해서 나올 수도 있지 않을까요?
+- [ ] Enum 활용
+- [ ] static한 클래스/로직이 나오면 이를 도메인 객체 내부로 옮길 수 있지 않을가 고민
+- [ ] name을 외부에서 주입하지 않고 내부에서 만들어도 되지 않을까 싶어요.
