+++ src/main/kotlin/blackjack/model/Player.kt
@@ -1,3 +1,3 @@
 package blackjack.model
 
-class Player(val name: String, val cardDeck: CardDeck)
+class Player(val name: String, val cardHand: CardHand)
